+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 
+**0.27.0**
+
+Added method to specify the converter for statement parameters explicitly.
+
+```
+statement.setParameterConverter("paramName", converterInstance).query("paramName", value)
+```
+
 **0.26.6**
 
 maintenance release